Resolve Format Circular Dependencies
