+++ src/format/docx/format-docx.ts
@@ -9,7 +9,7 @@ import { kFilterParams } from "../../config/constants.ts";
 import { Format } from "../../config/types.ts";
 import { mergeConfigs } from "../../core/config.ts";
 import { formatResourcePath } from "../../core/resources.ts";
-import { createWordprocessorFormat } from "../formats.ts";
+import { createWordprocessorFormat } from "../formats-shared.ts";
 
 const kIconCaution = "icon-caution";
 const kIconImportant = "icon-important";
