+++ src/format/formats.ts
@@ -8,50 +8,15 @@
 import { mergeConfigs } from "../core/config.ts";
 
 import {
-  kCache,
-  kCodeFold,
-  kCodeLink,
-  kCodeSummary,
-  kCodeTools,
   kEcho,
-  kError,
-  kEval,
-  kExecuteDaemon,
-  kExecuteDaemonRestart,
-  kExecuteDebug,
-  kExecuteEnabled,
-  kExecuteIpynb,
-  kFigAlign,
-  kFigDpi,
   kFigFormat,
   kFigHeight,
   kFigWidth,
-  kFreeze,
-  kInclude,
-  kIncludeInHeader,
-  kKeepHidden,
-  kKeepIpynb,
-  kKeepMd,
-  kKeepSource,
-  kKeepTex,
   kKeepYaml,
-  kLatexAutoInstall,
-  kLatexAutoMk,
-  kLatexClean,
-  kLatexMakeIndex,
-  kLatexMakeIndexOpts,
-  kLatexMaxRuns,
-  kLatexOutputDir,
-  kLatexTlmgrOpts,
-  kMergeIncludes,
-  kOutput,
   kOutputDivs,
-  kOutputExt,
   kOutputFile,
   kPageWidth,
   kPreferHtml,
-  kSelfContainedMath,
-  kStandalone,
   kVariant,
   kWarning,
 } from "../config/constants.ts";
@@ -62,7 +27,12 @@ import { htmlFormat } from "./html/format-html.ts";
 import { beamerFormat, latexFormat, pdfFormat } from "./pdf/format-pdf.ts";
 import { epubFormat } from "./epub/format-epub.ts";
 import { docxFormat } from "./docx/format-docx.ts";
-import { formatResourcePath } from "../core/resources.ts";
+import {
+  createEbookFormat,
+  createFormat,
+  createHtmlFormat,
+  createWordprocessorFormat,
+} from "./formats-shared.ts";
 
 export function defaultWriterFormat(to: string): Format {
   // to can sometimes have a variant, don't include that in the lookup here
@@ -213,63 +183,6 @@ export function defaultWriterFormat(to: string): Format {
   return writerFormat;
 }
 
-export function createFormat(ext: string, ...formats: Array<unknown>): Format {
-  return mergeConfigs(
-    defaultFormat(),
-    ...formats,
-    {
-      render: {
-        [kOutputExt]: ext,
-      },
-    },
-  );
-}
-
-export function createHtmlFormat(
-  figwidth: number,
-  figheight: number,
-) {
-  return createFormat("html", {
-    execute: {
-      [kFigFormat]: "retina",
-      [kFigWidth]: figwidth,
-      [kFigHeight]: figheight,
-    },
-    pandoc: {
-      [kStandalone]: true,
-    },
-  });
-}
-
-export function createEbookFormat(ext: string): Format {
-  return createFormat(ext, {
-    formatExtras: () => {
-      return {
-        [kIncludeInHeader]: [formatResourcePath("epub", "styles.html")],
-      };
-    },
-    render: {
-      [kMergeIncludes]: false,
-    },
-    execute: {
-      [kFigWidth]: 5,
-      [kFigHeight]: 4,
-    },
-  });
-}
-
-export function createWordprocessorFormat(ext: string): Format {
-  return createFormat(ext, {
-    render: {
-      [kPageWidth]: 6.5,
-    },
-    execute: {
-      [kFigWidth]: 5,
-      [kFigHeight]: 4,
-    },
-  });
-}
-
 function htmlPresentationFormat(figwidth: number, figheight: number): Format {
   return mergeConfigs(
     createHtmlFormat(figwidth, figheight),
@@ -343,59 +256,4 @@ function plaintextFormat(ext: string): Format {
       standalone: true,
     },
   });
-}
-
-function defaultFormat(): Format {
-  return {
-    execute: {
-      [kFigWidth]: 7,
-      [kFigHeight]: 5,
-      [kFigFormat]: "png",
-      [kFigDpi]: 96,
-      [kError]: false,
-      [kEval]: true,
-      [kCache]: null,
-      [kFreeze]: false,
-      [kEcho]: true,
-      [kOutput]: true,
-      [kWarning]: true,
-      [kInclude]: true,
-      [kKeepMd]: false,
-      [kKeepIpynb]: false,
-      [kExecuteIpynb]: null,
-      [kExecuteEnabled]: null,
-      [kExecuteDaemon]: null,
-      [kExecuteDaemonRestart]: false,
-      [kExecuteDebug]: false,
-    },
-    render: {
-      [kKeepTex]: false,
-      [kKeepYaml]: false,
-      [kKeepSource]: false,
-      [kKeepHidden]: false,
-      [kPreferHtml]: false,
-      [kOutputDivs]: true,
-      [kOutputExt]: "html",
-      [kFigAlign]: "default",
-      [kCodeFold]: "none",
-      [kCodeSummary]: "Code",
-      [kCodeLink]: false,
-      [kCodeTools]: false,
-      [kMergeIncludes]: true,
-      [kSelfContainedMath]: false,
-      [kLatexAutoMk]: true,
-      [kLatexAutoInstall]: true,
-      [kLatexClean]: true,
-      [kLatexMaxRuns]: 1,
-      [kLatexMaxRuns]: 10,
-      [kLatexMakeIndex]: "makeindex",
-      [kLatexMakeIndexOpts]: [],
-      [kLatexTlmgrOpts]: [],
-      [kLatexOutputDir]: null,
-    },
-    pandoc: {
-      from: "markdown",
-    },
-    metadata: {},
-  };
 }
